Do not use SHA lookup and use artifact list instead
