+++ external-providers/java-external-provider/pkg/java_external_provider/service_client.go
@@ -353,6 +353,14 @@ func (p *javaServiceClient) initialization(ctx context.Context) {
 
 }
 
+func (p *javaServiceClient) SetDepLabels(depLabels map[string]*depLabelItem) {
+	p.depToLabels = depLabels
+}
+
+func (p *javaServiceClient) SetExcludeDepLabels(depLabels map[string]*depLabelItem) {
+	p.depToLabels = depLabels
+}
+
 type openSourceLabels bool
 
 func (o openSourceLabels) GetLabels() []string {
